Added Plugin Support
The Validator can now be extended with custom validation plugins. See
the jQuery.Validator.EmailPlugin.js for an example. That plugin can also
be found in jQueryValidatorExtend.html (serves as an example of how to
build and use plugins). Custom Checks CAN be removed, but I'm letting
them stay for legacy support.